pw reset done
The pw reset works fine but you have to change the "max_execution_time" in the php.ini to min 1800 (30min). Cause it is not good to let a php file run for 30 min, i'm write to pw reset different now. I only pushed this one, in case we need it one day. sry for bad english i dont care, im tired, and im stressed, so shuttttt!!!